Move remove_file method from base service class.
Method is only used by pipeline and credentials service classes.
