+++ mash/services/mash_service.py
@@ -222,15 +222,6 @@ class MashService(object):
 
         return config['job_file']
 
-    def remove_file(self, config_file):
-        """
-        Remove file from disk if it exists.
-        """
-        try:
-            os.remove(config_file)
-        except Exception:
-            pass
-
     def restart_jobs(self, callback):
         """
         Restart jobs from config files.
